make readArrayFile a private function
